pyproject.toml
[project]
name = "asr"
version = "0.2.0"
description = "CrisperWhisper transcription optimized for Apple Silicon M4"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "mlx-whisper>=0.4.0",
    "huggingface_hub[hf_xet]>=0.20",  # Fast model downloads
    "pydantic>=2.0",
    "typer>=0.9",
    "rich>=13.0",
    "toml>=0.10",
    "anthropic>=0.30",
    "numpy>=1.24",
    "torchaudio>=2.0",  # For Silero VAD
    "python-dotenv>=1.0",  # Load .env files
    "jellyfish>=1.0",  # Phonetic matching (Soundex fallback)
    "rapidfuzz>=3.0",  # Fast fuzzy matching (56% faster than jellyfish)
    "Metaphone>=0.6",  # Double Metaphone support
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "ruff>=0.1",
]
ner = [
    "gliner>=0.2.20",  # Zero-shot NER for proper noun discovery
    "spacy>=3.7",  # Production NER
    "thinc-apple-ops>=1.0",  # Apple Silicon optimization
]
g2p = [
    "g2p-en>=2.1.0",  # Grapheme-to-phoneme for pronunciations
]

[project.scripts]
asr = "asr.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/asr"]

[tool.ruff]
line-length = 100
target-version = "py311"
